+++ forge/src/main/java/argent_matter/gcys/common/data/forge/GCySKeyMappingsImpl.java
@@ -13,7 +13,7 @@ public class GCySKeyMappingsImpl {
 
     }
 
-    @SubscribeEvent
+    //@SubscribeEvent
     public static void event(RegisterKeyMappingsEvent event) {
         event.register(GCySKeyMappings.START_ROCKET);
     }