+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 0.0.2.a:
 - fix fabric
+- make dyson sphere able to collapse if left without maintenance
+- add dyson sphe
+- remove rocket launch keybind
 
 
 0.0.2: